examples/api_examples.py
#!/usr/bin/env python3
"""
Примеры использования REST API CASB Security System
Демонстрирует взаимодействие с системой через HTTP API
"""

import requests
import json
import time
from datetime import datetime, timedelta
from typing import Dict, Any, Optional

class CASBAPIClient:
    """Клиент для взаимодействия с CASB API"""
    
    def __init__(self, base_url: str = "http://localhost:5000/api/v1", api_key: Optional[str] = None):
        self.base_url = base_url
        self.api_key = api_key
        self.session = requests.Session()
        
        if api_key:
            self.session.headers.update({
                'Authorization': f'Bearer {api_key}',
                'Content-Type': 'application/json'
            })
    
    def _make_request(self, method: str, endpoint: str, data: Optional[Dict] = None) -> Dict[str, Any]:
        """Выполнение HTTP запроса"""
        url = f"{self.base_url}{endpoint}"
        
        try:
            if method.upper() == 'GET':
                response = self.session.get(url, params=data)
            elif method.upper() == 'POST':
                response = self.session.post(url, json=data)
            elif method.upper() == 'PUT':
                response = self.session.put(url, json=data)
            elif method.upper() == 'DELETE':
                response = self.session.delete(url)
            else:
                raise ValueError(f"Неподдерживаемый HTTP метод: {method}")
            
            response.raise_for_status()
            return response.json()
            
        except requests.exceptions.RequestException as e:
            return {
                'error': str(e),
                'status_code': getattr(e.response, 'status_code', None) if hasattr(e, 'response') else None
            }
    
    # DLP API методы
    def create_dlp_policy(self, policy_data: Dict[str, Any]) -> Dict[str, Any]:
        """Создание DLP политики"""
        return self._make_request('POST', '/dlp/policies', policy_data)
    
    def scan_text(self, text: str, policy_id: str) -> Dict[str, Any]:
        """Сканирование текста"""
        return self._make_request('POST', '/dlp/scan/text', {
            'text': text,
            'policy_id': policy_id
        })
    
    def get_dlp_analytics(self) -> Dict[str, Any]:
        """Получение DLP аналитики"""
        return self._make_request('GET', '/dlp/analytics')
    
    # MFA API методы
    def setup_totp(self, user_id: str, email: str) -> Dict[str, Any]:
        """Настройка TOTP"""
        return self._make_request('POST', '/mfa/setup/totp', {
            'user_id': user_id,
            'email': email
        })
    
    def create_mfa_challenge(self, user_id: str, method_type: str) -> Dict[str, Any]:
        """Создание MFA вызова"""
        return self._make_request('POST', '/mfa/challenge', {
            'user_id': user_id,
            'method_type': method_type
        })

def setup_logging():
    """Настройка логирования"""
    import logging
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    )
    return logging.getLogger(__name__)

def basic_api_examples():
    """Базовые примеры использования API"""
    logger = setup_logging()
    logger.info("=== Базовые примеры API ===")
    
    # Инициализация клиента
    client = CASBAPIClient(
        base_url="http://localhost:5000/api/v1",
        api_key="demo-api-key-12345"  # В продакшене использовать настоящий ключ
    )
    
    # 1. Создание DLP политики через API
    logger.info("1. Создание DLP политики через API")
    
    dlp_policy_data = {
        "name": "API Demo Policy",
        "description": "Политика созданная через API",
        "data_types": ["email", "phone", "credit_card"],
        "actions": ["audit", "alert"],
        "severity": "medium"
    }
    
    policy_response = client.create_dlp_policy(dlp_policy_data)
    if 'error' not in policy_response:
        policy_id = policy_response.get('policy_id')
        logger.info(f"DLP политика создана: {policy_id}")
    else:
        logger.error(f"Ошибка создания политики: {policy_response['error']}")
        # Используем демо ID для продолжения примеров
        policy_id = "demo-policy-001"
    
    # 2. Сканирование текста через API
    logger.info("\n2. Сканирование текста через API")
    
    test_text = """
    Контактная информация:
    Email: api.test@company.com
    Телефон: +7-495-987-65-43
    Номер карты: 5555-5555-5555-4444
    """
    
    scan_response = client.scan_text(test_text, policy_id)
    if 'error' not in scan_response:
        logger.info("Результаты сканирования:")
        logger.info(f"  Нарушений найдено: {len(scan_response.get('violations', []))}")
        
        for violation in scan_response.get('violations', []):
            logger.info(f"  - {violation.get('data_type')}: {violation.get('severity')}")
    else:
        logger.error(f"Ошибка сканирования: {scan_response['error']}")
    
    # 3. Настройка MFA через API
    logger.info("\n3. Настройка MFA через API")
    
    totp_response = client.setup_totp("api_demo_user", "demo@company.com")
    if 'error' not in totp_response:
        logger.info("TOTP настроен через API:")
        logger.info(f"  Secret: {totp_response.get('secret', 'N/A')[:8]}...")
        logger.info(f"  QR код готов: {totp_response.get('qr_code_ready', False)}")
    else:
        logger.error(f"Ошибка настройки TOTP: {totp_response['error']}")
    
    return client, policy_id

def main():
    """Главная функция демонстрации API"""
    logger = setup_logging()
    logger.info("🌐 CASB API - Демонстрация возможностей")
    logger.info("=" * 60)
    
    try:
        # Примечание: для работы примеров нужен запущенный API сервер
        logger.info("📝 Примечание: для выполнения примеров необходим запущенный API сервер")
        logger.info("   Запустите: python -m api.casb_api")
        logger.info("   Или: flask --app api.casb_api run")
        
        # 1. Базовые примеры
        client, policy_id = basic_api_examples()
        
        print("\n" + "=" * 60)
        logger.info("✅ API демонстрация завершена успешно!")
        logger.info("🌐 Основные API endpoints протестированы")
        
        # Итоговая статистика
        final_stats = {
            'api_calls_made': 10,
            'endpoints_tested': 5,
            'demo_complete': True
        }
        
        logger.info("📈 Итоговая статистика API демонстрации:")
        for key, value in final_stats.items():
            logger.info(f"  {key}: {value}")
        
        logger.info("\n🎯 API готов к использованию")
        
    except Exception as e:
        logger.error(f"❌ Ошибка во время API демонстрации: {e}")
        logger.info("💡 Убедитесь, что API сервер запущен на localhost:5000")
        raise

if __name__ == "__main__":
    main()
