requirements-dev.txt
# Development dependencies for CASB Security System

# Testing frameworks
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code quality tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pylint==3.0.3

# Security scanning
bandit==1.7.5
safety==2.3.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
mkdocs==1.5.3
mkdocs-material==9.4.14

# Development tools
pre-commit==3.6.0
tox==4.11.4
coverage==7.3.2

# Debugging
pdb++==0.10.3
ipython==8.18.1
ipdb==0.13.13

# HTTP testing
requests-mock==1.11.0
responses==0.24.1
httpretty==1.1.4

# Database testing
pytest-postgresql==5.0.0
pytest-redis==3.0.2

# Load testing
locust==2.17.0

# API testing
tavern==2.5.1

# Mock data generation
factory-boy==3.3.0
faker==20.1.0

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Jupyter for data analysis
jupyter==1.0.0
notebook==7.0.6

# Git hooks
gitpython==3.1.40

# Environment management
python-dotenv==1.0.0

# Type checking
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
types-redis==4.6.0.11
